+++ angular/app/unsupported_browser/unsupported_browser.js
@@ -1,0 +1,8 @@
+(function(){
+    "use strict";
+
+    angular.module('app.controllers').controller('UnsupportedBrowserCtrl', function(){
+        //
+    });
+
+})();